+++ src/test/java/com/openggf/game/rewind/TestScalarOnlyCodecDeletion.java
@@ -580,6 +580,30 @@ public class TestScalarOnlyCodecDeletion {
                             "com.openggf.game.sonic2.objects.SmallMetalPformObjectInstance$SmallMetalPformChildInstance",
                             "xFlipped"));
 
+    private static final List<MutableFieldCoverageCandidate> BATCH136_MUTABLE_FIELDS =
+            List.of(
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.AizBombExplosionInstance",
+                            "posY"),
+                    new MutableFieldCoverageCandidate(
+                            AizCollapsingLogBridgeObjectInstance.class.getName(),
+                            "hFlip", "halfWidth", "isFireBridge", "subtypeBase", "totalTimer", "x", "y"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.AizCollapsingLogBridgeObjectInstance$CollapsingLogSegment",
+                            "fixedX", "isFireVariant"),
+                    new MutableFieldCoverageCandidate(
+                            AizDisappearingFloorObjectInstance.class.getName(),
+                            "periodMask", "phaseOffset", "x", "y"),
+                    new MutableFieldCoverageCandidate(
+                            "com.openggf.game.sonic3k.objects.AizDisappearingFloorObjectInstance$BorderChild",
+                            "x", "y"),
+                    new MutableFieldCoverageCandidate(
+                            AizFallingLogObjectInstance.class.getName(),
+                            "phaseOffset", "spawnX", "spawnY", "timingMask"),
+                    new MutableFieldCoverageCandidate(
+                            AizFlippingBridgeObjectInstance.class.getName(),
+                            "animDirection", "animPeriod", "hFlip", "maxFrame", "x", "y"));
+
     private static final List<CodecDeletionCandidate> BATCH31_DELETED_CODECS = List.of(
             new CodecDeletionCandidate(BombPrizeObjectInstance.class.getName(), GameId.S2));
 
@@ -3794,6 +3818,23 @@ public class TestScalarOnlyCodecDeletion {
     @Test
     void batch135S2LauncherConstructorScalarsAreMutableForCompactRestore() {
         for (MutableFieldCoverageCandidate candidate : BATCH135_MUTABLE_FIELDS) {
+            Class<?> cls = loadClass(candidate.fqn());
+            for (String fieldName : candidate.fieldNames()) {
+                try {
+                    var field = findField(cls, fieldName);
+                    assertFalse(Modifier.isFinal(field.getModifiers()),
+                            cls.getName() + "#" + fieldName
+                                    + " must be mutable so compact restore can replay captured scalars");
+                } catch (NoSuchFieldException e) {
+                    throw new AssertionError("Missing scalar field " + cls.getName() + "#" + fieldName, e);
+                }
+            }
+        }
+    }
+
+    @Test
+    void batch136S3kAizConstructorScalarsAreMutableForCompactRestore() {
+        for (MutableFieldCoverageCandidate candidate : BATCH136_MUTABLE_FIELDS) {
             Class<?> cls = loadClass(candidate.fqn());
             for (String fieldName : candidate.fieldNames()) {
                 try {